Remove lr test
Summary:
The name of the blob holding the learning rate is determined programatically and can change based off of where it's being run. Here we have it as "SgdOptimizer_0_lr_cpu", but it fails for me because that isn't the name of the blob storing the learning rate.

The blob is created here: https://fburl.com/7wkv2kp8, https://fburl.com/8ivhp7o5

To reverse engineer the name from the code above seems a little hacky. We could instead make the lr blob's name accessible by storing it as a member of SGDOptimizer, but that change would be within the caffe2 codebase and would therefore require that users of our open source repo install caffe2 from its open source. This seems like a lot of trouble for little gain; this test was really only testing functionality limited to caffe2's SGDOptimizer, not our code.

My weak inclination is to just remove this test, but I'm happy to make the change in the caffe2 code instead if that seems preferable.

Reviewed By: MisterTea

Differential Revision: D6494743

fbshipit-source-id: 01c326f70f2541052284cdc0daa4d7429aa817a0