Overhaul normalization code
Summary:
This is an overhaul of the normalization code:
(1) The preprocessed matrix is now sorted by feature type.
(2) All feature types except enum are processed in batch.
(3) Using the built-in boxcox op.
(4) The exporter code has been rewritten to be modular.
(5) the continuous model & discrete model exporters are now small skeletons that call into the parent class RLPredictor.
(6) Boxcox is now a separate feature type

Differential Revision: D6935416

fbshipit-source-id: 0937ba94404abc54985a0ed9b4a1d51fc55c7762