Speed up RL discrete action training, increase CPU worker threads / remove
Summary:
Profiling using prof_dag revealed there is some sequential code that can be parallelized.
Increased the number of workers in different net defs and removed the NanCheck operator from
the preprocessor net as the CPU version of the operator is very slow.
Preprocess enum and quantile features in smaller batches while training.

Reviewed By: econti

Differential Revision: D8072428

fbshipit-source-id: 1680428f579e61d608eac7db05a0165cf7669abd