Create GridworldSamples container class
Summary:
GridWorld::generate_samples(...) returns a batch of preprocessed training examples as a tuple of batches. As we add components to this set, the tuple has grown prohibitively large.

This task is to replace this tuple with a class GridworldSamples that will hold this information so we do not need to return so many arguments.

Reviewed By: MisterTea

Differential Revision: D8509223

fbshipit-source-id: 4a1f662158022349531d50f6430a04d51d3356c2