weighted doubly robust
Summary:
Sequential doubly-robust estimator is unbiased, but of high-variance, partly due to the high variance of importance weights. Unbiasness is a desirable property, however, it may not be what we really aim for. In the end, we try to build an estimator that predicts the Q-values that are close to the true values if we indeed run the target policy. In another words, we try to minimize the MSE between the true values and our estimators.

It is a very common technique to utilize weighted importance weights to decrease the variance, at the expense of introducing bias. Hopefully by doing this tradeoff, we can reduce MSE.

In this diff, we incorporate weighted importance weights into our sequential doubly-robust estimator.

Please find more details at https://arxiv.org/pdf/1604.00923.pdf Section 5.

Some preliminary results on Cartpole

If using direct method to compute V-estimates:

{F132962838}

{F132962840}

{F132962841}

If not using direct method to compute V-estimates (pick the max of the Q-estimates):

{F132963629}

{F132963628}

{F132963630}

Using direct method has higher bias but smaller variance. On the other hand, not using direct method would give more accurate estimates for most of the time (smaller bias), but when the estimates are not accurate, they are generally off by a larger margin (larger variance).

Overall, the weighted version has a smaller variance compared to the ordinary doubly robust estimator.

Note the true values in our setting should be the discounted true P.E instead of the ordinary true P.E.

Reviewed By: MisterTea

Differential Revision: D8874222

fbshipit-source-id: feadaa26d0b64d692d268e59fd5fa408c5dd4973