MAGIC estimator
Summary:
Model and Guided Importance Sampling Combining (MAGIC) Estimator

For details, please visit https://arxiv.org/pdf/1604.00923.pdf Section 7, 8

For a doubly robust estimator, its terms can be decomposed as
     per_decision_importance_sampled_rewards (PDIS) + direct_method_estimates (DM) - combined control variate (which has a mean of 0)

The difference between this estimator and weighted doubly robust estimator,
    the weighted DR simply combines PDIS and DM together by adding summing them up.
    on the other side, the new estimator blends them together. To be more specific, PDIS is used up to time_step j, and DM is used after step j.

For each j, we have a estimated value and we can it j-step return

The main motivation, is that PDIS is unbiased but of high variance. On contrast, DM is biased but of relatively smaller variance.

So the new estimator sets up a framework to balance the two by having a small sets of j and corresponding j_step returns.

Since our ultimate goal is to minimize the loss between (true value - our estimate). This estimator directly use the ultimate error (of course we do not know the true loss, but we can approximate it by doing bias+variance decomposition), to choose the right normalized weights for the j_step returns.

Preliminary results on cartpole

{F133197993}

The screenshot is from one run. After running for 10 times, the mean of absolute error between the new estimator and the true discounted value is 0.1522 with a range of 0.374 and standard deviation of 0.13501

Reviewed By: econti

Differential Revision: D8907140

fbshipit-source-id: 3820d1429647b3904254f77a3c1234d02780fcbe