Added tests for CPE and some supporting functions in the gridworld environment
Summary: Both the logged and model policies are epsilon-greedy policies where greedy = optimal, but the epsilon values are different. We test a variety of epsilon pairs to check the estimator's ability to evaluate model policies that are much different than the logged policies that generated the data. By computing the true values associated with both epsilon policies, we can see the performance and compute a percentage error.

Reviewed By: econti

Differential Revision: D13765911

fbshipit-source-id: 558da1a147e676f4f7ee8d4362ab49427fdd4166