Sequence feature extractor
Summary:
Implement reading ID-list & ID-score-list schema into SequenceFeatures. Sequence truncation & padding is implemented. If the sequence is longer than max-length, the front of the sequence is truncated. If the sequence is shorter than max-length, the back of the sequence is padded with 0.

Id-mapping will be coming in the next diff.

Reviewed By: MisterTea

Differential Revision: D14382970

fbshipit-source-id: e1e5817d0995657bfa16fd694510227f599b762c