Enable SAC without value network + Fix GPU for gym + Fix SAC with value network
Summary: Modified SAC to include a hyperparameter which selects whether or not to use the value network (use Q(s,pi(s) in target instead). Corrects bug which causes SAC to fail on gym tasks with GPU. Corrects bug which causes SAC to use batch action in target update of V, rather than the policy action.

Reviewed By: econti

Differential Revision: D15327627

fbshipit-source-id: 8dbb114380079be5ede34cff2cc8579220c8029b