Add another simple POMDP
Summary:
Add another simple POMDP which I call String Game. In this game, the agent can observe a character at one time. But the
reward is given based on last n (n>2) steps' observation (a string). The transition is simple too: the action taken is equal to the next state.

Reviewed By: kittipatv

Differential Revision: D15480139

fbshipit-source-id: 7dddde5711f917cc7934e12b7d06423ad2a523e3