refactor horizon tests under a base class
Summary: When multiple tests are run in the same process (as in OSS CI), we need to make sure that SummaryWriterContext is reset for each test. Previously, this required explicitly writing ``setUp()`` and ``tearDown()`` methods for every test class to reset SummaryWriterContext. Now, the test classes inherit from a base test class (``HorizonTestBase``) which handles resetting SummaryWriterContext.

Reviewed By: kittipatv

Differential Revision: D16165688

fbshipit-source-id: 5b38edf4f2b7a7a9ac0ce79256fac08c9491f697