MountainCar PBT experiment workflows
Summary: We want to compare the relative effectiveness of PBT vs. Bayesian Optimization for the automated tuning of reward shaping weights. Although it will take some time to see results from the online testing on stories notifications, we can run preliminary experiments on a toy problem - due to the sparsity of its reward signal, we are using MountainCar. Intermediate dense rewards are defined by binning/discretizing the position/velocity of the agent and assigning a reward proportional to the magnitude of the agent's velocity. This is the diff for the PBT experiment workflow.

Reviewed By: kittipatv

Differential Revision: D16998608

fbshipit-source-id: 3936a3419ee918c0296ac84a17c7cf43c79e9ec2