Rl: Dqn-Preprocessing
Summary:
Buckification of Spark projects: This diff is forRl: Dqn-preprocessing app. Using it pom.xml file as a guide, a TARGET file was created for the app.

Background Info:
For all Java projects in fbcode, build will be done with buck. Since mid-2018, there's an in-flight move to get completely onto buck for development and deployment of all internal Java code: See:
https://fb.facebook.com/groups/java.eng/permalink/1626137397434869/

Reviewed By: k21

Differential Revision: D20428514

fbshipit-source-id: cf9cfaaa90e56ccc9b7323a3c6e62b3d71f4a1e8