Change seq2slate reward network architecture
Summary:
Previously we have a reward model that uses only items shown in the actual slate to predict the slate reward.
{F233991957}

I think it is better to let the reward model know more information from other candidates. So I modify the reward model to have a very similar architecture as the full seq2slate:

{F233993799}

Reviewed By: kittipatv

Differential Revision: D21011830

fbshipit-source-id: 40efc002b83f05d5f73a87f32b26b3409ba116bd