Changed the way we're integrating sequential OPE Estimators to be more dynamic
Summary: Instead of "simulating" the mdps by their indices, I've changed the evaluation data page to have an RLEstimatorInput if it's being used in the sequential setting. This RLEstimatorInput exposes the actual target policy and estimated value function to the OPE module, instead of precomputing the propensities/values before evaluation.

Reviewed By: jia-git

Differential Revision: D22826052

fbshipit-source-id: fb428fb84583c06948e7cd99db2821993633bba9