Add iterative-softmax decoder for Seq2Slate (#316)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/316

We used to only have an autoregressive decoder for seq2slate. In this diff we add an iterative-softmax decoder for seq2slate. It only computes encoder scores once and then perform iterative softmax at each step until the full slate is decoded. It is equivalent to Frechet Policy Gradient mathematically.

We also add and fix more tests in the diff.

Reviewed By: badrinarayan

Differential Revision: D23807407

fbshipit-source-id: f11d997f7fadddf472942cf746be6a72f0c6d670