Fix LR schedulers in ReAgent + supporting tests
Summary:
Fixed LR schedulers in ReAgent by adding a "verbose" field to each to match their torch.optim counterparts.

To test this, a new `optimizer_tests` target has been added to ensure that a valid Optimizer can be created for each type of LR scheduler (excluding LambdaLR, MultiplicativeLR, and CyclicLR for now as they are not supported yet).

Differential Revision: D24841422

fbshipit-source-id: 86285a2c13e21b832402d1f1b11a364d23d6c786