Fix LR schedulers in ReAgent + supporting tests (#344)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/344

Fixed LR schedulers in ReAgent by adding a "verbose" field to each to match their torch.optim counterparts.

To test this, a new `optimizer_tests` target has been added to ensure that a valid Optimizer can be created for each type of LR scheduler (excluding LambdaLR, MultiplicativeLR, and CyclicLR for now as they are not supported yet).

Reviewed By: alexnikulkov

Differential Revision: D24841422

fbshipit-source-id: 6be13a92d7206aa693b0068f10744aeb3d4efb4a