Computing numbers needed for CFEval in validation_step (#460)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/460

OOM issues can occur in CFEval of DQN and CRR workflows when the validation set is too large, as in https://fb.workplace.com/groups/horizon.users/permalink/836921400197015/. This diff solves this issue by computing the numbers needed for CFEval in `validation_step`, instead of just stacking the raw batches, which include all the state features that can take a lot of memory.

Note that if `use_gpu=True`, for speed the CFEval-required numbers are computed on the GPUs, where both the validation batch and the trainer is stored. Then the returned `EvaluationDataPage` will be moved to the CPU, because later in `validation_epoch_end` everything will be done on the CPU for larger memory capacity. To enable this transportation between devices, in this diff `EvaluationDataPage` is changed to a subclass of `TensorDataClass` from the previous `NamedTuple`.

Reviewed By: kaiwenw

Differential Revision: D27929283

fbshipit-source-id: 07928a95faf6c77184a1709c3e5897552168c021