Add discount_time_scale to SlateQTrainer
Summary:
According to the original [SlateQ paper](https://arxiv.org/abs/1905.12767) (p28, 2nd paragraph, last sentence), the discount factor `gamma` will be scaled by the time difference in this way:

`gamma^((t2-t1)/time_scale)`.

Here, `t1` and `t2` are the timestamps between the current and the next state-action pairs within a training sample, and the `time_scale` is a hyperparameter that can scale up/down the time difference.

This diff implements this mechanism by adding a `discount_time_scale` parameter to `SlateQTrainer`. Its value is the `time_scale` in the formula above.

If this parameter is not set, i.e., `None`, we will keep the discount factor as it is.

Reviewed By: kittipatv

Differential Revision: D29297804

fbshipit-source-id: 5bd9101a2fe3b1b3d9817a3233357cab197e8ce8