Reducing log noise
Summary: There are several modules in the ReAgent library where the logger level is set in the library code thus overriding the level set by the library client resulting in very verbose stdout. This diff removes places in the library where the logger level is set so that the client's setting is always maintained.

Reviewed By: bankawas

Differential Revision: D29673661

fbshipit-source-id: 8f6db342571d4524768f75d6d6bf4416bad8ad1c