Migrating from fvcore to iopath
Summary:
Our current PathManager is based on fvcore, it's in the process of deprecation and is being replaced by an open source solution iopath.

This diff is the result of running the provided codemod script on reagent, rl, and rl_exp, followed by a round of autodeps and a fbgs search for 'fvcore'.

https://fb.workplace.com/groups/939200583171018/permalink/1022439911513751/

Reviewed By: czxttkl

Differential Revision: D29974786

fbshipit-source-id: 397fd69ef94d43a7ca07c963c2a46bbbdcf78599