add option for configuring slate_size in multi selection (#517)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/517

If we don't enable `single_selection`, we should find a way to calculate the next slate value from all Q-values of all items on the next slate.

The default way to calculate that is by summing up all the next Q-values and average by the slate_size of the next slate.

There is another way to average by the current slate_size as well.

This wasn't an issue before since the slate_size is fixed, but after we landed D29848923 (https://github.com/facebookresearch/ReAgent/commit/5351f6313cb24fbd9ff604beaae99b66ee9eed0e), the slate size can be different.

We're not sure theoretically if either averaging method is better, so we propose this diff to allow configuring that.

The new option is called `next_slate_value_norm_method` and it can take:

- `"norm_by_current_slate_size"`: sum the next slate Q-values and average by the **current** slate size;
- `"norm_by_next_slate_size"`: sum the next slate Q-values and average by the **next** slate size;

cc achechetka solofsson

Reviewed By: czxttkl

Differential Revision: D29986728

fbshipit-source-id: f178b5da1462e4f9cc6995367ed229ab958c477a