Add foreach flag to reagent optimizer configs
Summary:
A new foreach flag is being added to the optimizers to indicate whether foreach logic or single tensor logic is used (see D33767870).

This causes reagent tests to fail such as https://www.internalfb.com/intern/testinfra/diagnostics/7318349469673867.281475021413633.1644559942/

The issue arises from this line https://fburl.com/code/lroy3a2p where the value for foreach cannot be found in `getattr(self, k)`.

This PR adds the foreach flag to `uninferrable_optimizers.py` to address this (Note that we do not add this flag to `LBFGS` and `SparseAdam` as they do not support this option)

Differential Revision: D34216723

fbshipit-source-id: c0456329590056404ab50b845c1c7ed3c8008c4a