Add foreach flag to reagent optimizer configs (#606)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/606

A new foreach flag is being added to the optimizers to indicate whether foreach logic or single tensor logic is used (see D33767870 and the associated stack).

This causes reagent tests to fail such as https://www.internalfb.com/intern/testinfra/diagnostics/7318349469673867.281475021413633.1644559942/

The issue arises from this line https://fburl.com/code/lroy3a2p where the value for foreach cannot be found in `getattr(self, k)`.

This PR adds the foreach flag to `uninferrable_optimizers.py` to address this (Note that we do not add this flag to `LBFGS` and `SparseAdam` as they do not support this option)

Reviewed By: alexnikulkov

Differential Revision: D34216723

fbshipit-source-id: fac4e6095157c7cd33184bfa5b7042bdd151688e