Refactor sparse arch and interaction arch (#112)
Summary:
X-link: https://github.com/pytorch/torchrec/pull/112

As discussed in D33960410, we want the responsibility of processing KeyedTensor into sparse features the responsibility of SparseArch.

A motivation for this is that we want to have an extension EsuhmDLRM, where all we would need to do is replace the sparse arch component. However, the esuhm sparse arch's output doesn't adhere to the current KeyedTensor output.

Reviewed By: bigning

Differential Revision: D34482853

fbshipit-source-id: f38e2a07bce67b746e686d9a5d7bd593986359e3