docs: add GH button in support of Ukraine (#613)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/613

Reviewed By: alexnikulkov

Differential Revision: D34630775

Pulled By: dmitryvinn-fb

fbshipit-source-id: 7108199313663725759377fe0972e59e9ae2cb22