Add check if the logger is set in dqn_trainer (#650)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/650

In the base LightningModule class we define an optional `logger` property which may be
initialized to None, while in DQNtrainer._log_dqn method we try to access the `logger`
object without checking first if it was initialized. The issue surfaced when trying to run
unit tests analogous to those in `test_qrdqn`. This commit adds a check whether
the `logger` is initialized prior to attempting to use it.
Interestingly, the analogous QRDQNTrainer class implementation does not use the
`logger` property for logging, perhaps it's redundant?

Differential Revision: D37529027

fbshipit-source-id: 9c79e17dd0c929703b39c8ce311591799bb6f0d6