improve LinUCB on-demand coefficient calculation (#661)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/661

Make `coefs` a property, instead of an attribute. This property automatically checks if the currently saved coefficient value is valid (computed using the current values of `A` and `b`) and uses it if so. If the current saved coefficient value is invalid, we compute the new value as `coefs = A**(-1) * b` and save it

Reviewed By: czxttkl

Differential Revision: D38283370

fbshipit-source-id: aa78828e6a8cf8a7971ccee734a2cbe71856ccd1