change LinearRegressionUCB forward pass logic to make it more tracing-friendly
Summary:
The "if" statement in the forward pass causes a warning during tracing. It will be safer to use a separate `_forward_no_coefs_check()` method for the forward pass of a trained model. In this method we don't check the validity of coefficients, instead the coefficients are assumed to be valid because we compute them at the end of the training epoch

The behavior of existing methods/properties should remain unchanged.

Differential Revision: D39692331

fbshipit-source-id: bbc4e9a1c50721f46a3cd210c42001a0871bdecd