change LinearRegressionUCB forward pass logic to make it more tracing-friendly (#682)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/682

The "if" statement in the forward pass causes a warning during tracing. It will be safer to use a separate `_forward_no_coefs_check()` method for the forward pass of a trained model. In this method we don't check the validity of coefficients, instead the coefficients are assumed to be valid because we compute them at the end of the training epoch

The behavior of existing methods/properties should remain unchanged.

Reviewed By: czxttkl

Differential Revision: D39692331

fbshipit-source-id: bcb2a705345ad34e7ab1574c21955dd111267de9