Fix Multi-processor in LinUCB training
Summary:
1. We train LinUCB models on two hours (7am and 8am) separately and get covariance matrixes A1 and A2 for each.
2. We train a LinUCB model at 7am, enable recurring so it launched a recurring training at 8am. Then we get covariance matrix A.

We found that A is around 8*A1 + A2 instead of A1 + A2, where we used NPROC_PER_NODE = 8 when training.

The reason is the following:

Current implementation aggregates A and b calculated in each training processor. However, if we do recurring training, then in the 2nd training child, it will aggregate the previous A NPROC_PER_NODE times and then add the new A aggregated from current epoch.

This diff fixes the above issue.

Reviewed By: alexnikulkov

Differential Revision: D39902734

fbshipit-source-id: 73a9e9525e68d75e2dd558b75e1db8a3bea7255a