Fix reagent LRScheduler tests
Summary:
I recently exposed LRScheduler as a public endpoint as that is the right direction for users. This diff adds LRScheduler as a torch lr scheduler, which it is.

Would fix test errors such as https://www.internalfb.com/intern/testinfra/diagnostics/281475249445597.562950030008200.1668085134/, which were introduced by my landing of D41109279.

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: czxttkl

Differential Revision: D41187073

fbshipit-source-id: 659b3fa6d3d97c5b94151f302b330e1702418cd8