Fix reagent LRScheduler tests (#697)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/697

I recently exposed LRScheduler as a public endpoint as that is the right direction for users. This diff adds LRScheduler as a torch lr scheduler, which it is.

Would fix test errors such as https://www.internalfb.com/intern/testinfra/diagnostics/281475249445597.562950030008200.1668085134/, which were introduced by my landing of D41109279.

Created from CodeHub with https://fburl.com/edit-in-codehub

Reviewed By: czxttkl

Differential Revision: D41187073

fbshipit-source-id: 2637b6a80247c24620cf0ce8310e8181135637cd