Track average A and b instead of cumulative in LinUCB (#707)
Summary:
Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/707

I'm updating the training logic of LinUCB to keep track of the average values of `A` and `b` instead of cumulative values. This should improve the numerical stability of training by preventing numerical overflows.

The average values are aggregated among the trainers and among epochs when computing the coefficients.

Reviewed By: BerenLuthien

Differential Revision: D42334470

fbshipit-source-id: f103c9ab2fe84fa2da639d7ec726c82737d6c4f0