update lighting version specification in OSS reagent (#711)
Summary:
The current version is specified via a commit hash. This started generating an error during install for an unknown reason. I analyzed the releases and the commit history - 1.6.0 is the closest released version to this commit. I tried using a recent lighting version, but it generated quite a few errors on CircleCI due to changes in APIs.

Pull Request resolved: https://github.com/facebookresearch/ReAgent/pull/711

Reviewed By: speedystream, BillMatrix

Differential Revision: D42873869

fbshipit-source-id: fd7b34fc7193cfb15f6d4ff5d9d379e7c422835d