refactor LinUCBTrainer to remove the wrapper
Summary:
Moving the functionality from the wrapper straight to the trainer. This will be useful to support hydra-based trainer instantiation.
To enable this I also updated metric logging: instead of logging to the summary_writer, now we're using the standard Lighting interface `self.logger.log_metrics()`

NOTE: Because of the difference in logging frequency of supervised and Offline Eval metrics the plots for Offline Eval have flat regions where the same value is logged several times. This is a mostly harmless quirk and it shouldn't break any existing offline eval workflows.

Reviewed By: BerenLuthien

Differential Revision: D43590414

fbshipit-source-id: d5d82959890d82bc7d1505ad7d4b878ada33f020