Add support for label transformations to Contextual Bandits in FREE
Summary:
Adding a separate concept of `label`, which is used as the prediction target for model training. In the basic case, `label` is equal to `reward` and if `CBInput` is created without specifying `label`, we automatically set `label=reward` in `__post_init__`. But we can also define `label` differently, e.g. as a transformation of `reward`, to give the model a more stable learning target.
I have observed improvements in performance from using `log` or `sqrt` transforms in AP Container Selection.
The `reward` field is now used only for Offline Evaluation, while `label` is used for model training and supervised learning accuracy metrics.
In a FREE workflow the transform is specified in `config.features.label_transform`, which can be one of `["identity", "log", "sqrt"]`

Reviewed By: BerenLuthien, PoojaAg18

Differential Revision: D45300610

fbshipit-source-id: 7005b10e652549948e9104c9d90ef76475276f67