Append column of ones to MLP output in Neural LinUCB to account for bias/intercept
Summary: Currently we pass the output of MLP directly to the input of LinUCB. This is equivalent to using a linear layer without a bias term. This diff appends a column of ones as an extra feature to the output of MLP in order to allow LinUCB to have a bias/intercept.

Reviewed By: PoojaAg18

Differential Revision: D45539068

fbshipit-source-id: 48504c276590c04e274b380f89f806b2307809cf