Add importance weights to Policy Evaluator
Summary:
Currently observations from low-cardinality campaigns have a higher probability of being accepted. This skews the reward estimator to focus more on low-cardinality campaigns. We need to apply importance weights to normalize the estimator. The importance weight for each observations is 1/logged_prob (In the case of uniform exploration the probability is equal to `1/CampaignSize`). The weight is used both for training and to weight the reward in the Offline Eval.

The Importance Weights can be effectively disabled by setting `max_importance_weight=1.0`

Changes made in:
1. reagent CB batch preprocessor to extract the logged action probability(+ test)
2. CB eval utils + test
3. FREE AP data utils

Reviewed By: BerenLuthien

Differential Revision: D46411892

fbshipit-source-id: d39282259b4961d03d66314f6046bb5231692d7b